refactor: App.vue の整理と useAppProps の削除
